requirements.txt
selenium==3.141.0
pytest==6.1.1
requests==2.25.1
pytest-xdist==2.2.1
allure-pytest==2.8.40